fix controllers add lombok mappin order fix entity

+++ src/main/java/com/softlines/fastpos/controller/CategoryController.java
@@ -1,13 +1,17 @@
 package com.softlines.fastpos.controller;
 
 import com.softlines.fastpos.domain.Category;
+import com.softlines.fastpos.domain.Product;
+import com.softlines.fastpos.dto.CategoryDto;
+import com.softlines.fastpos.dto.ProductDto;
+import com.softlines.fastpos.dto.mapping.CategoryMapper;
+import com.softlines.fastpos.dto.service.DtoService;
 import com.softlines.fastpos.repository.CategoryRepository;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.util.Assert;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
-
 import java.util.List;
 import java.util.Optional;
 
@@ -15,42 +19,43 @@ import java.util.Optional;
 @RequestMapping("/category")
 public class CategoryController {
 
+    @Autowired
     private CategoryRepository categoryRepository;
-
-    public CategoryController(CategoryRepository categoryRepository) {
-        this.categoryRepository = categoryRepository;
-    }
+    @Autowired
+    CategoryMapper categoryMapper;
+    @Autowired
+    DtoService dtoService;
 
     @PostMapping("/save")
-    public ResponseEntity<Category> addCategory(@RequestBody Category category) {
+    public ResponseEntity<CategoryDto> addCategory(@RequestBody CategoryDto categoryDto) {
         try {
-            Category existingCategory = categoryRepository.findById(category.getId()).get();
-            if (existingCategory != null) {
-                return ResponseEntity.status(HttpStatus.NO_CONTENT).body(existingCategory);
-            }
-            Category createdCategory = categoryRepository.save(category);
-            if (existingCategory != null) {
-                return ResponseEntity.status(HttpStatus.NO_CONTENT).body(category);
+            Optional<Category> optionalCategory = categoryRepository.findById(categoryDto.getId());
+
+            if (!optionalCategory.isPresent()) {
+                Category createdCategory = categoryRepository.save(dtoService.categoryDtoToCategory(categoryDto));
+
+                return ResponseEntity.status(HttpStatus.CREATED).body(categoryMapper.toCategoryDto( createdCategory));
             } else {
-                return ResponseEntity.status(HttpStatus.CREATED).body(createdCategory);
+                return ResponseEntity.notFound().build();
 
             }
         } catch (Exception exception) {
             throw new ResponseStatusException(
+
                     HttpStatus.NOT_FOUND, " Not Found", exception);
         }
     }
 
     @GetMapping("/getall")
-    public ResponseEntity<List<Category>> getCategories() {
+    public ResponseEntity<List<CategoryDto>> getCategories() {
         try {
 
-
-            if (categoryRepository.findAll() == null) {
-                return ResponseEntity.status(HttpStatus.NO_CONTENT).body(categoryRepository.findAll());
+            List<Category> categories = categoryRepository.findAll();
+            if (categories == null) {
+                return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
 
             } else {
-                return ResponseEntity.ok().body(categoryRepository.findAll());
+                return ResponseEntity.ok().body(categoryMapper.toCategoryDTOs(categories));
 
             }
         } catch (Exception exception) {
@@ -61,13 +66,13 @@ public class CategoryController {
     }
 
     @GetMapping("/get/{id}")
-    public ResponseEntity<Optional<Category>> getCategory(@PathVariable long id) {
+    public ResponseEntity<CategoryDto> getCategory(@PathVariable long id) {
         try {
-            if (categoryRepository.findById(id) == null) {
-                return ResponseEntity.status(HttpStatus.NO_CONTENT).body(categoryRepository.findById(id));
-
+            Optional<Category> optionalCategory = categoryRepository.findById(id);
+            if (optionalCategory.isPresent()) {
+                return ResponseEntity.status(HttpStatus.NO_CONTENT).body(categoryMapper.toCategoryDto(optionalCategory.get()));
             } else {
-                return ResponseEntity.ok().body(categoryRepository.findById(id));
+                return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
             }
         } catch (Exception exception) {
             throw new ResponseStatusException(
@@ -75,26 +80,55 @@ public class CategoryController {
         }
     }
 
-    @PutMapping("/put/{id}")
-    public void editCategory(@PathVariable long id, @RequestBody Category category) {
+    @GetMapping("/getByName/{name}")
+    public ResponseEntity<List<CategoryDto>> getProductByName(@PathVariable String name) {
         try {
-            Category existingCategory = categoryRepository.findById(id).get();
-            Assert.notNull(existingCategory, "Category not found");
-            existingCategory.setBackgroundString(existingCategory.getBackgroundString());
-            existingCategory.setRank(existingCategory.getRank());
-            existingCategory.setName(existingCategory.getName());
-            categoryRepository.save(existingCategory);
+
+            List<Category> categories = categoryRepository.findByName(name);
+
+            if (categories != null)
+                return ResponseEntity.ok(categoryMapper.toCategoryDTOs(categories));
+            else
+                return ResponseEntity.notFound().build();
+
         } catch (Exception exception) {
             throw new ResponseStatusException(
                     HttpStatus.NOT_FOUND, " Not Found", exception);
         }
     }
 
+    @PutMapping("/put/{id}")
+    public ResponseEntity<CategoryDto> editCategory(@PathVariable long id, @RequestBody CategoryDto categoryDto) {
+        try {
+
+            Optional<Category> optionalCategory = categoryRepository.findById(id);
+
+            if (optionalCategory.isPresent()) {
+                return ResponseEntity.ok().body(categoryMapper.toCategoryDto(categoryRepository.save(dtoService.categoryDtoToCategory(categoryDto))));
+            } else {
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+            }
+        } catch (Exception exception) {
+            throw new ResponseStatusException(
+                    HttpStatus.NOT_FOUND, " Not Found", exception);
+        }
+
+    }
+
     @DeleteMapping("/delete/{id}")
-    public void deleteCategory(@PathVariable long id) {
+    public ResponseEntity deleteCategory(@PathVariable long id) {
         try {
             Category categoryToDel = categoryRepository.findById(id).get();
-            categoryRepository.delete(categoryToDel);
+
+            if (categoryToDel != null) {
+
+                categoryRepository.delete(categoryToDel);
+                return ResponseEntity.ok().build();
+
+            } else {
+                return ResponseEntity.notFound().build();
+            }
+
         } catch (Exception exception) {
             throw new ResponseStatusException(
                     HttpStatus.NOT_FOUND, " Not Found", exception);
